change spring/damping calculation, added centre of mass option
The spring/damper value will now be calculatet based on absolut values, means when you change the spring length, the spring/damp stiffnes will be the same.

when u work with normalized values, the spring/damp stiffness will change when you change the spring length

added a node "CentreOfMass".
it will change the com, based on the nodes position, so you can change the COM without changing the collider size/pos.
could manipulate it to represent the fuel consume